Refactor users controller
