+++ app/controllers/api/users_controller.rb
@@ -2,17 +2,6 @@
 
 module Api
   class UsersController < ApplicationController
-    skip_before_action :user_profile_filled, only: %i[show update]
-
-    def show
-      render json: UserSerializer.render(current_user, view: :full)
-    end
-
-    def update
-      current_user.update!(user_params)
-      render json: UserSerializer.render(current_user, view: :full)
-    end
-
     def follow
       current_user.following_users << user
       render status: :no_content
@@ -24,10 +13,6 @@ module Api
     end
 
     private
-
-    def user_params
-      params.require(:user).permit(%i[name bio website email birthdate password username])
-    end
 
     def user
       User.find_by!(username: params[:username])
